convert integer to float64 using float64()

+++ app.go
@@ -7,5 +7,5 @@ func main() {
 	var expectedReturnRate = 5.5 
 	var years = 10
 
-	var futureValue = investmentAmount * (1 + expectedReturnRate / 100)
+	var futureValue = float64(investmentAmount) * (1 + expectedReturnRate / 100)
 }
